requirements.txt
[Abstract fetching]
biopython-1.85
numpy-2.2.5
pandas-2.2.3 (python-dateutil-2.9.0.post0 pytz-2025.2 six-1.17.0 tzdata-2025.2)
tqdm-4.67.1

[OpenAI API NER-RE]
openai-1.77.0 (annotated-types-0.7.0 anyio-4.9.0 certifi-2025.4.26 distro-1.9.0 h11-0.16.0 httpcore-1.0.9 httpx-0.28.1 idna-3.10 jiter-0.9.0 pydantic-2.11.4 pydantic-core-2.33.2 sniffio-1.3.1 typing-extensions-4.13.2 typing-inspection-0.4.0)

[bioBERT]
torch-2.7.0
transformers-4.51.3 (MarkupSafe-3.0.2 charset-normalizer-3.4.2 filelock-3.18.0 fsspec-2025.3.2 huggingface-hub-0.30.2 jinja2-3.1.6 mpmath-1.3.0 networkx-3.4.2 nvidia-cublas-cu12-12.6.4.1 nvidia-cuda-cupti-cu12-12.6.80 nvidia-cuda-nvrtc-cu12-12.6.77 nvidia-cuda-runtime-cu12-12.6.77 nvidia-cudnn-cu12-9.5.1.17 nvidia-cufft-cu12-11.3.0.4 nvidia-cufile-cu12-1.11.1.6 nvidia-curand-cu12-10.3.7.77 nvidia-cusolver-cu12-11.7.1.2 nvidia-cusparse-cu12-12.5.4.2 nvidia-cusparselt-cu12-0.6.3 nvidia-nccl-cu12-2.26.2 nvidia-nvjitlink-cu12-12.6.85 nvidia-nvtx-cu12-12.6.77 packaging-25.0 pyyaml-6.0.2 regex-2024.11.6 requests-2.32.3 safetensors-0.5.3 sympy-1.14.0 tokenizers-0.21.1 triton-3.3.0 urllib3-2.4.0)

[Requests]
beautifulsoup4-4.13.4 soupsieve-2.7
lxml-5.4.0

[BERN2 QC]
matplotlib-3.10.3 (contourpy-1.3.2 cycler-0.12.1 fonttools-4.58.0 kiwisolver-1.4.8 pillow-11.2.1 pyparsing-3.2.3)